reorganise project files
- Move APi models into own folder
- Move providers
- Move repository
