--- app-backend/backend/Controllers/DateTimeProvider.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace backend.Controllers;
-
-public class DateTimeProvider
-{
-    public DateTime UtcNow() => 
-        GetUtcNow();
-
-    protected virtual DateTime GetUtcNow() => 
-        DateTime.UtcNow;
-}
