--- app-backend/backend/Controllers/TransactionIdProvider.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace backend.Controllers;
-
-public class TransactionIdProvider
-{
-    public Guid NewTransactionId() => 
-        GetTransactionId();
-
-    protected virtual Guid GetTransactionId() => 
-        Guid.NewGuid();
-}
