+++ app-backend/backend/Controllers/TransactionsController.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using backend.Models;
+using backend.Repository;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
