+++ app-backend/backend/Models/TransactionRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace backend.Controllers;
+namespace backend.Models;
 
 public class TransactionRequest
 {
