+++ app-backend/backend/Providers/DateTimeProvider.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace backend.Providers;
+
+public class DateTimeProvider
+{
+    public DateTime UtcNow() =>
+        GetUtcNow();
+
+    protected virtual DateTime GetUtcNow() =>
+        DateTime.UtcNow;
+}
