+++ app-backend/backend/Providers/TransactionIdProvider.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace backend.Providers;
+
+public class TransactionIdProvider
+{
+    public Guid NewTransactionId() =>
+        GetTransactionId();
+
+    protected virtual Guid GetTransactionId() =>
+        Guid.NewGuid();
+}
